Added schedule checking and database saving
Database saving will break when HTML form is updated with different
input types. Will fix when we come to that point. For now it works as a
first test.

Input form now looks at scheduleData.js and makes the appropriate list
of teams available to select for the currently selected match number.